+++ pkg/kudoctl/util/kudo/kudo_test.go
@@ -23,20 +23,6 @@ func newTestSimpleK2o() *Client {
 	return NewClientFromK8s(fake.NewSimpleClientset(), kubefake.NewSimpleClientset())
 }
 
-func TestKudoClientValidate(t *testing.T) {
-	tests := []struct {
-		err string
-	}{
-		{"failed to run crd verification: failed to retrieve CRD"}, // verify that NewClient tries to validate CRDs
-	}
-
-	for _, tt := range tests {
-		_, err := NewClient("testdata/test-config", 0, true)
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), tt.err)
-	}
-}
-
 func TestKudoClient_OperatorExistsInCluster(t *testing.T) {
 
 	obj := v1beta1.Operator{
