--- pkg/kudoctl/util/kudo/testdata/test-config
@@ -1,14 +1,0 @@
-clusters:
-- cluster:
-    server: 127.0.0.1:1
-  name: cluster
-contexts:
-- context:
-    cluster: cluster
-    user: user
-  name: cluster
-current-context: cluster
-preferences: {}
-users:
-- name: user
-  user: {}
